# docs/swagger.yaml
basePath: /
definitions:
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/todo.TodoList'
        type: array
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  todo.ListsItem:
    properties:
      id:
        type: integer
      itemId:
        type: integer
      listId:
        type: integer
    type: object
  todo.TodoList:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  todo.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  utility.ErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for TodoList Application
  title: Todo App API
  version: "1.0"
paths:
  /api/lists:
    get:
      consumes:
      - application/json
      description: get all lists
      operationId: get-all-lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: create todo list
      operationId: create-list
      parameters:
      - description: list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.TodoList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create todo list
      tags:
      - lists
  /api/lists/:id:
    get:
      consumes:
      - application/json
      description: get list by id
      operationId: get-list-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.ListsItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get List By Id
      tags:
      - lists
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"